+++ packages/wallet/src/json-rpc-validation/schema/create-channel.json
