+++ packages/wallet/src/json-rpc-validation/schema/definitions.json
