+++ packages/wallet/src/json-rpc-validation/schema/get-address.json
