+++ packages/wallet/src/json-rpc-validation/schema/join-channel.json
