+++ packages/wallet/src/json-rpc-validation/schema/notification.json
