+++ packages/wallet/src/json-rpc-validation/schema/push-message.json
