+++ packages/wallet/src/json-rpc-validation/schema/request.json
