+++ packages/wallet/src/json-rpc-validation/schema/response.json
